+++ CHANGELOG.md
@@ -9,7 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- ...
+- Overlay: Detailed speed readout, with raw readout option (setting: `show_speed` `show_speed_raw`)
+- Overlay: Terrain speed offsets readout (setting: `show_speed_offsets`)
+- Overlay: Simplified FPS readout option (setting: `show_fps_simple`)
+- Overlay: MFG (bounce glitch) timer (setting: `show_mfg_timer`)
+
+### Changed
+
+- Overlay: Removed general practice mode restriction; practice mode now defined for individual components
 
 ## [0.1.6] - 2024-08-11